# openapi.json
{"openapi":"3.0.2","info":{"title":"Swap","description":"A Simple Api for finding good paths between two token swaps (/path) and create a non singed transaction which can then be signed via a wallet singer (meamask , web3 <- private key , ...) ","version":"1.1"},"paths":{"/log/":{"get":{"summary":"Get Log","operationId":"get_log_log__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/swap":{"post":{"summary":"Swap","description":"Swap method :\n    Inorder to create a transaction data while calling path method to find best path and distributions between two tokens ! :)\n    This function creates transaction to 1inch router !\n    Last Step would be another application signing this transation with a (from address ) wallet.\n      ","operationId":"swap_swap_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Item"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/path":{"post":{"summary":"Path","description":"path method :\n    finds good path and destributions from token to token :)","operationId":"path_path_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Item"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/toturial-metamask":{"get":{"summary":"Sign With Metamsak","description":"Visit <a href=\"http://localhost:8000/toturial-metamask\">Here</a>","operationId":"sign_with_metamsak_toturial_metamask_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/toturial-nodejs":{"get":{"summary":"Sign With Metamsak","description":"Visit <a href=\"http://localhost:8000/toturial-nodejs\">Here</a>","operationId":"sign_with_metamsak_toturial_nodejs_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Item":{"title":"Item","required":["fromToken","toToken"],"type":"object","properties":{"walletAddress":{"title":"Walletaddress","type":"string"},"fromToken":{"title":"Fromtoken","type":"string"},"toToken":{"title":"Totoken","type":"string"},"amount":{"title":"Amount","type":"number"},"slippage":{"title":"Slippage","type":"integer"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}